pug-or-ugh/requirements.txt
coverage==4.5.4
Django==2.2.4
djangorestframework==3.10.2
pytz==2019.2
sqlparse==0.3.0
